Pushing logic into helper file
