+++ tests/test_sql_refsols/tpch_q6_sqlite.sql
@@ -1,35 +1,39 @@
 SELECT
-  SUM(TEMP_COL0) AS REVENUE
+  COALESCE(agg_0, 0) AS REVENUE
 FROM (
   SELECT
-    L_EXTENDEDPRICE * L_DISCOUNT AS TEMP_COL0
+    SUM(amt) AS agg_0
   FROM (
     SELECT
-      L_DISCOUNT,
-      L_EXTENDEDPRICE
+      extended_price * discount AS amt
     FROM (
       SELECT
-        L_DISCOUNT,
-        L_EXTENDEDPRICE,
-        L_QUANTITY,
-        L_SHIPDATE
-      FROM LINEITEM
+        discount,
+        extended_price
+      FROM (
+        SELECT
+          l_discount AS discount,
+          l_extendedprice AS extended_price,
+          l_quantity AS quantity,
+          l_shipdate AS ship_date
+        FROM tpch.LINEITEM
+      )
+      WHERE
+        (
+          discount <= 0.07
+        )
+        AND (
+          quantity < 24
+        )
+        AND (
+          ship_date < '1995-01-01'
+        )
+        AND (
+          discount >= 0.05
+        )
+        AND (
+          ship_date >= '1994-01-01'
+        )
     )
-    WHERE
-      (
-        L_QUANTITY < 24
-      )
-      AND (
-        L_DISCOUNT <= 0.07
-      )
-      AND (
-        L_DISCOUNT >= 0.05
-      )
-      AND (
-        L_SHIPDATE < '1995-01-01'
-      )
-      AND (
-        L_SHIPDATE >= '1994-01-01'
-      )
   )
 )
